change signed to unsigned

+++ process.c
@@ -139,16 +139,16 @@ void sendbinary(FILE* file, int sock){
 	rewind(file);
 
 	int send_buff_size = size;
-	long *send_buff = NULL;
-	send_buff = malloc(send_buff_size*sizeof(long));
+	uint32_t *send_buff = NULL;
+	send_buff = malloc(send_buff_size*sizeof(uint32_t));
 
 	assert(send_buff != NULL);
 
 	// Read from the file (two bytes at a time)
 	int i = 0, j = 0;
 	long just_read;
-	for(; j < size; i++, j += sizeof(long)){
-		fread(&just_read, sizeof(long), 1, file);
+	for(; j < size; i++, j += sizeof(uint32_t)){
+		fread(&just_read, sizeof(uint32_t), 1, file);
 		send_buff[i] = htonl(just_read);
 	}
